requirements.txt
Django==3.0
django-rest-framework==0.1.0
websockets==8.1
graphene-django==2.15.0
celery==5.0.5
redis==3.5.3
